attempt at ensuring the fft is done on the actual recorded data
